board: moved player's code into separated file
